# optimize insertFrom of ColumnAggregateFunction to share Aggregate State in some cases