--- a/tests/performance/bitmap_array_element.xml
+++ b/tests/performance/bitmap_array_element.xml
@@ -1,0 +1,12 @@
+<test>
+    <query>
+		WITH
+            (
+                SELECT bitmapBuild(groupArray(number))
+                FROM numbers(10000000)
+            ) AS a,
+            [a, a, a] AS b
+        SELECT sum(bitmapCardinality(b[(number % 3) + 1]))
+        FROM numbers(10000)
+    </query>
+</test>