--- a/tests/performance/bitmap_array_element.xml
+++ b/tests/performance/bitmap_array_element.xml
@@ -1,6 +1,6 @@
 <test>
     <query>
-		WITH
+        WITH
             (
                 SELECT bitmapBuild(groupArray(number))
                 FROM numbers(10000000)