--- a/tests/performance/bitmap_array_element.xml
+++ b/tests/performance/bitmap_array_element.xml
@@ -3,7 +3,7 @@
         WITH
             (
                 SELECT bitmapBuild(groupArray(number))
-                FROM numbers(10000000)
+                FROM numbers(3000000)
             ) AS a,
             [a, a, a] AS b
         SELECT sum(bitmapCardinality(b[(number % 3) + 1]))