### Changelog category (leave one):

- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
optimize insertFrom of ColumnAggregateFunction to share Aggregate State in some cases.

`insertFrom` method of ColumnAggregateFunction always `ensureOwnership` of Aggregation State and make a copy for 
new inserted element, but in some cases  it's unneed and state can be shared. Like below:
- before optimize:
```sql
vccliu1645086827254-0 :) WITH
                             (
                                 SELECT bitmapBuild(groupArray(number))
                                 FROM numbers(10000000)
                             ) AS a,
                             [a, a, a] AS b
                         SELECT sum(bitmapCardinality(b[(number % 3) + 1]))
                         FROM numbers(1000)

WITH
    (
        SELECT bitmapBuild(groupArray(number))
        FROM numbers(10000000)
    ) AS a,
    [a, a, a] AS b
SELECT sum(bitmapCardinality(b[(number % 3) + 1]))
FROM numbers(1000)

Query id: 51861099-7acf-4e81-b7a1-75d97dfe62fc

┌─sum(bitmapCardinality(arrayElement(b, plus(modulo(number, 3), 1))))─┐
│                                                         10000000000 │
└─────────────────────────────────────────────────────────────────────┘

1 row in set. Elapsed: 1.253 sec. Processed 10.02 million rows, 80.19 MB (8.00 million rows/s., 64.02 MB/s.)
```
- after optimize:
```sql
vccliu1645086827254-0 :) WITH
                             (
                                 SELECT bitmapBuild(groupArray(number))
                                 FROM numbers(10000000)
                             ) AS a,
                             [a, a, a] AS b
                         SELECT sum(bitmapCardinality(b[(number % 3) + 1]))
                         FROM numbers(1000)

WITH
    (
        SELECT bitmapBuild(groupArray(number))
        FROM numbers(10000000)
    ) AS a,
    [a, a, a] AS b
SELECT sum(bitmapCardinality(b[(number % 3) + 1]))
FROM numbers(1000)

Query id: dc58f763-6e46-4f4d-ac72-e4f81ab6b9ff

┌─sum(bitmapCardinality(arrayElement(b, plus(modulo(number, 3), 1))))─┐
│                                                         10000000000 │
└─────────────────────────────────────────────────────────────────────┘

1 row in set. Elapsed: 0.460 sec. Processed 10.02 million rows, 80.19 MB (21.77 million rows/s., 174.14 MB/s.)
```
The main causes here is `b[number % 3 + 1]` which will call `arrayElement -> ColumnAggregateFunction::insertFrom`, before it make a lot of large bitmap copy but now all shared.