@CurtizJ 2022-10-03
Test failures are unrelated.

```
Some query of the test 'bitmap_array_element' is too slow to run. See the all queries report

The test 'bitmap_array_element' is too slow to run as a whole. Investigate whether the create and fill queries can be sped up
```

I'll reduce the size of input in perf test.